+++ banan_image.cpp
@@ -18,15 +18,15 @@
 #include <ImfArray.h>
 
 namespace Banan {
-    BananImage::BananImage(BananDevice &device, uint32_t width, uint32_t height, uint32_t mipLevelsCount, uint32_t arrayLevelsCount, VkFormat format, VkImageTiling tiling, VkSampleCountFlagBits numSamples, VkImageUsageFlags usageFlags, VkMemoryPropertyFlags memoryPropertyFlags) : bananDevice{device}, imageFormat{format}, mipLevels{mipLevelsCount}, arrayLevels{arrayLevelsCount} {
+    BananImage::BananImage(BananDevice &device, size_t width, size_t height, size_t mipLevelsCount, size_t arrayLevelsCount, VkFormat format, VkImageTiling tiling, VkSampleCountFlagBits numSamples, VkImageUsageFlags usageFlags, VkMemoryPropertyFlags memoryPropertyFlags) : bananDevice{device}, imageFormat{format}, mipLevels{mipLevelsCount}, arrayLevels{arrayLevelsCount} {
         VkImageCreateInfo imageInfo{};
         imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
         imageInfo.imageType = VK_IMAGE_TYPE_2D;
         imageInfo.extent.width = static_cast<uint32_t>(width);
         imageInfo.extent.height = static_cast<uint32_t>(height);
         imageInfo.extent.depth = 1;
-        imageInfo.mipLevels = mipLevels;
-        imageInfo.arrayLayers = arrayLevels;
+        imageInfo.mipLevels = static_cast<uint32_t>(mipLevels);
+        imageInfo.arrayLayers = static_cast<uint32_t>(arrayLevels);
         imageInfo.format = imageFormat;
         imageInfo.tiling = tiling;
         imageInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
@@ -35,7 +35,7 @@ namespace Banan {
         imageInfo.samples = numSamples;
 
         bananDevice.createImageWithInfo(imageInfo, memoryPropertyFlags, image, memory);
-        imageExtent = {width, height};
+        imageExtent = {static_cast<uint32_t>(width), static_cast<uint32_t>(height)};
         imageLayout = VK_IMAGE_LAYOUT_UNDEFINED;
 
         createTextureSampler();
@@ -69,9 +69,9 @@ namespace Banan {
         viewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D;
         viewInfo.format = imageFormat;
         viewInfo.subresourceRange.baseMipLevel = 0;
-        viewInfo.subresourceRange.levelCount = mipLevels;
+        viewInfo.subresourceRange.levelCount = static_cast<uint32_t>(mipLevels);
         viewInfo.subresourceRange.baseArrayLayer = 0;
-        viewInfo.subresourceRange.layerCount = arrayLevels;
+        viewInfo.subresourceRange.layerCount = static_cast<uint32_t>(arrayLevels);
 
         VkFormat depthFormats[] = {VK_FORMAT_D32_SFLOAT_S8_UINT, VK_FORMAT_D32_SFLOAT, VK_FORMAT_D24_UNORM_S8_UINT, VK_FORMAT_D16_UNORM_S8_UINT, VK_FORMAT_D16_UNORM};
         if (std::find(std::begin(depthFormats), std::end(depthFormats), imageFormat) != std::end(depthFormats)) {
@@ -92,9 +92,9 @@ namespace Banan {
         viewInfo.viewType = VK_IMAGE_VIEW_TYPE_2D_ARRAY;
         viewInfo.format = imageFormat;
         viewInfo.subresourceRange.baseMipLevel = 0;
-        viewInfo.subresourceRange.levelCount = mipLevels;
+        viewInfo.subresourceRange.levelCount = static_cast<uint32_t>(mipLevels);
         viewInfo.subresourceRange.baseArrayLayer = 0;
-        viewInfo.subresourceRange.layerCount = arrayLevels;
+        viewInfo.subresourceRange.layerCount = static_cast<uint32_t>(arrayLevels);
 
         VkFormat depthFormats[] = {VK_FORMAT_D32_SFLOAT_S8_UINT, VK_FORMAT_D32_SFLOAT, VK_FORMAT_D24_UNORM_S8_UINT, VK_FORMAT_D16_UNORM_S8_UINT, VK_FORMAT_D16_UNORM};
         if (std::find(std::begin(depthFormats), std::end(depthFormats), imageFormat) != std::end(depthFormats)) {
@@ -141,7 +141,7 @@ namespace Banan {
     }
 
     void BananImage::generateMipMaps(uint32_t m) {
-        bananDevice.generateMipMaps(image, imageExtent.width, imageExtent.height, m);
+        bananDevice.generateMipMaps(image, static_cast<size_t>(imageExtent.width), static_cast<size_t>(imageExtent.height), m);
         mipLevels = m;
     }
 
@@ -257,14 +257,14 @@ namespace Banan {
         return bananImage;
     }
 
-    BananCubemap::BananCubemap(BananDevice &device, uint32_t sideLength, uint32_t mipLevelsCount, VkFormat format, VkImageTiling tiling, VkSampleCountFlagBits numSamples, VkBufferUsageFlags usageFlags, VkMemoryPropertyFlags memoryPropertyFlags) : bananDevice{device}, cubemapImageFormat{format}, mipLevels{mipLevelsCount} {
+    BananCubemap::BananCubemap(BananDevice &device, size_t sideLength, size_t mipLevelsCount, VkFormat format, VkImageTiling tiling, VkSampleCountFlagBits numSamples, VkBufferUsageFlags usageFlags, VkMemoryPropertyFlags memoryPropertyFlags) : bananDevice{device}, cubemapImageFormat{format}, mipLevels{mipLevelsCount} {
         VkImageCreateInfo imageInfo{};
         imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
         imageInfo.imageType = VK_IMAGE_TYPE_2D;
         imageInfo.extent.width = static_cast<uint32_t>(sideLength);
         imageInfo.extent.height = static_cast<uint32_t>(sideLength);
         imageInfo.extent.depth = 1;
-        imageInfo.mipLevels = mipLevels;
+        imageInfo.mipLevels = static_cast<uint32_t>(mipLevels);
         imageInfo.arrayLayers = 6;
         imageInfo.format = cubemapImageFormat;
         imageInfo.tiling = tiling;
@@ -293,7 +293,7 @@ namespace Banan {
         imageViewCreateInfo.viewType = VK_IMAGE_VIEW_TYPE_CUBE;
         imageViewCreateInfo.format = cubemapImageFormat;
         imageViewCreateInfo.components = { VK_COMPONENT_SWIZZLE_R, VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY, VK_COMPONENT_SWIZZLE_IDENTITY };
-        imageViewCreateInfo.subresourceRange = { VK_IMAGE_ASPECT_COLOR_BIT, 0, mipLevels, 0, 6};
+        imageViewCreateInfo.subresourceRange = { VK_IMAGE_ASPECT_COLOR_BIT, 0, static_cast<uint32_t>(mipLevels), 0, 6};
         imageViewCreateInfo.image = cubemapImage;
 
         VkFormat depthFormats[] = {VK_FORMAT_D32_SFLOAT_S8_UINT, VK_FORMAT_D32_SFLOAT, VK_FORMAT_D24_UNORM_S8_UINT, VK_FORMAT_D16_UNORM_S8_UINT, VK_FORMAT_D16_UNORM};
